7-aug-2013

+++ app/database/migrations/2013_08_02_063249_create_org_table.php
@@ -25,7 +25,9 @@ class CreateOrgTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('org');
+		if (Schema::hasTable('org')) {
+			Schema::drop('org');
+		}
 	}
 
 }
